backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
python-multipart==0.0.9
openai
python-dotenv
stripe
